Add common stuff to common package
