+++ src/authentication/admin.py
@@ -1,24 +1,15 @@
-from django.contrib import admin
+from django.contrib.admin import StackedInline, register, site
 from django.contrib.auth.admin import GroupAdmin, UserAdmin
 from django.contrib.auth.forms import AdminPasswordChangeForm, UserChangeForm, UserCreationForm
 from django.contrib.auth.models import Group
-from django.forms import ValidationError, models
+from django.forms import ValidationError
+
+from common.admin import NoDeleteInline
 
 from .models import GroupExtension, User, UserProfile
 
 
-class NoDeleteInline(models.BaseInlineFormSet):
-    """
-    Custom formset to prevent deletion.
-    Used by the inline for userprofiles to prevent the possibility
-    of deleting the profile object.
-    """
-    def __init__(self, *args, **kwargs):
-        super(NoDeleteInline, self).__init__(*args, **kwargs)
-        self.can_delete = False
-
-
-class UserProfileInline(admin.StackedInline):
+class UserProfileInline(StackedInline):
     model = UserProfile
     formset = NoDeleteInline
 
@@ -38,9 +29,9 @@ class ImmutableAdminPasswordChangeForm(AdminPasswordChangeForm):
         raise ValidationError("Users can not be changed locally.")
 
 
-@admin.register(User)
+@register(User)
 class UserProfileAdmin(UserAdmin):
-    inlines = (UserProfileInline,)
+    inlines = [UserProfileInline]
     list_display = ("username", "email", "first_name", "last_name", "is_staff", "is_superuser", "is_active", "date_joined", "last_login")
     list_filter = ("groups", "is_staff", "is_superuser", "is_active")
     filter_horizontal = ("groups",)
@@ -49,15 +40,15 @@ class UserProfileAdmin(UserAdmin):
     change_password_form = ImmutableAdminPasswordChangeForm
 
 
-class GroupExtensionInline(admin.StackedInline):
+class GroupExtensionInline(StackedInline):
     model = GroupExtension
     formset = NoDeleteInline
 
 
-admin.site.unregister(Group)
+site.unregister(Group)
 
 
-@admin.register(Group)
+@register(Group)
 class GroupExtensionAdmin(GroupAdmin):
     inlines = (GroupExtensionInline,)
     search_fields = ("name",)
